🎨 [Structure]: #59 5 タスク CRUD ラッパを taskCommands.ts に集約
タスク CRUD は同じ関心事 (BE のタスク操作 invoke 群) なのに 1 関数 1 ファイル
で散らばっており凝集度が低かったため、`src/lib/tauri/taskCommands.ts` 1 本に
集約。各 Params 型も同居させて関連型と関数を 1 箇所で参照できるようにした。
テストファイルは `taskCommands.{operation}.test.ts` に改名し、対象名を
集約後のソースに合わせた。
