+++ src/lib/tauri/__tests__/taskCommands.deleteTask.test.ts
@@ -1,6 +1,6 @@
 import { invoke } from "@tauri-apps/api/core";
 import { beforeEach, expect, test, vi } from "vitest";
-import { deleteTask } from "@/lib/tauri/deleteTask";
+import { deleteTask } from "@/lib/tauri/taskCommands";
 import { TauriError } from "@/lib/tauri/tauriError";
 
 vi.mock("@tauri-apps/api/core", () => ({ invoke: vi.fn() }));
