--- src/lib/tauri/openProject.ts
@@ -1,28 +1,0 @@
-import type { Task } from "@/types/task";
-import type { Result } from "@/utils/result";
-import { invokeWrapped } from "./invokeWrapped";
-import type { TauriError } from "./tauriError";
-
-/** open_project 引数。 */
-export type OpenProjectParams = {
-  /** プロジェクトディレクトリの絶対パス */
-  path: string;
-};
-
-/** open_project 戻り値ペイロード（BE 仕様準拠）。 */
-export type OpenProjectPayload = {
-  /** プロジェクト内のタスク一覧 */
-  tasks: Task[];
-  /** カラム名一覧 */
-  columns: string[];
-};
-
-/**
- * プロジェクトディレクトリを開き、タスク・カラム名一覧を取得する。
- * @param params 引数オブジェクト
- * @returns 成功時は Result.ok(payload)、失敗時は Result.err(TauriError)
- */
-export const openProject = (
-  params: OpenProjectParams,
-): Promise<Result<OpenProjectPayload, TauriError>> =>
-  invokeWrapped<OpenProjectPayload>("open_project", params);
